requirements.txt
certifi==2024.7.4
elastic-transport==8.15.0
elasticsearch==8.15.0
paho-mqtt==2.1.0
urllib3==2.2.2
certifi==2024.7.4
charset-normalizer==3.3.2
elastic-transport==8.15.0
elasticsearch==8.15.0
Events==0.5
idna==3.7
opensearch-py==2.7.0
paho-mqtt==2.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
six==1.16.0
urllib3==2.2.2
